### 2022-05-05 #4379 @editorialbot
[PRE REVIEW]: unifir: A Unifying API for Working with Unity in R
**Submitting author:** @mikemahoney218 (<a href="http://orcid.org/0000-0003-2402-304X">Michael Mahoney</a>)
**Repository:** https://github.com/ropensci/unifir
**Branch with paper.md** (empty if default branch): paper
**Version:** v0.2.0
