### 2022-05-05 #4379 @kyleniemeyer
Hi @mikemahoney218, I see your submission was already accepted by rOpenSci. That's great! However, this may not quite meet our requirements for [research software](https://joss.readthedocs.io/en/latest/submitting.html#what-we-mean-by-research-software), so the editorial board is going to take a look. This should hopefully take about a week; any additional context/justification you can give would be helpful.
