### 2022-05-05 #4379 @mikemahoney218
Happily! I'd point to the section under "why do this": https://github.com/ropensci/unifir#why-do-this as well as to the realm of IVE research (papers from my group include our work with {terrainr} -- https://www.mm218.dev/papers/vrs_2021.pdf and https://joss.theoj.org/papers/10.21105/joss.04060 ; research from other groups includes https://link.springer.com/article/10.1007/s42489-020-00069-6 https://www.tandfonline.com/doi/full/10.1080/13658816.2020.1830997 https://www.sciencedirect.com/science/article/pii/S0143622818312207 https://www.sciencedirect.com/science/article/pii/S0095069608000983 , and a search on Google Scholar for "immersive research environment" + "Unity" yields 2,970 results). We view this as software to "support... the execution of research experiments" by enabling IVEs to be produced in an efficient and reproducible manner, hopefully improving the openness and reliability of the IVE literature and experiments conducted within IVEs as a result. Let me know if I can provide more information!
