### 2022-05-05 #4379 @mikemahoney218
I should also mention that I bring up {terrainr} because JOSS has already published software from our group that produces IVEs ( https://joss.theoj.org/papers/10.21105/joss.04060 ) -- {unifir} is a direct extension of that work, and we believe it qualifies as research software for the same reasons that {terrainr} did back in January.
