### 2022-05-06 #4379 @danielskatz
@mikemahoney218 - Please work on the possibly missing DOI and invalid DOIs that editorialbot suggests, but note that the missing one may be incorrect. Please feel free to make changes to your .bib file, then use the command `@editorialbot check references` to check again, and the command `@editorialbot generate pdf` when the references are right to make a new PDF. editorialbot commands need to be the first entry in a new comment.
