+++ paper.bib
@@ -1,7 +1,7 @@
 @inproceedings{Slater1993,
   author={Slater, M. and Usoh, M.},
-  booktitle={Proceedings of IEEE Virtual Reality Annual International Symposium}, 
-  title={Presence in immersive virtual environments}, 
+  booktitle={Proceedings of IEEE Virtual Reality Annual International Symposium},
+  title={Presence in immersive virtual environments},
   year={1993},
   volume={},
   number={},
@@ -37,6 +37,7 @@ doi = {10.1080/13658816.2020.1830997},
   volume          = {71},
   year            = {2021},
   pages           = {53-65},
+  doi = {10.1007/s42489-020-00069-6}
 }
 
   @Article{terrainr,
@@ -59,7 +60,7 @@ volume = {113},
 pages = {102102},
 year = {2019},
 issn = {0143-6228},
-doi = {https://doi.org/10.1016/j.apgeog.2019.102102},
+doi = {10.1016/j.apgeog.2019.102102},
 url = {https://www.sciencedirect.com/science/article/pii/S0143622818312207},
 author = {R.D. Swetnam and J. Korenko},
 }
@@ -73,7 +74,7 @@ pages = {65-86},
 year = {2009},
 note = {Frontiers of Environmental and Resource Economics},
 issn = {0095-0696},
-doi = {https://doi.org/10.1016/j.jeem.2008.08.002},
+doi = {10.1016/j.jeem.2008.08.002},
 url = {https://www.sciencedirect.com/science/article/pii/S0095069608000983},
 author = {Stephen M. Fiore and Glenn W. Harrison and Charles E. Hughes and E. Elisabet Rutström},
 }
