### 2022-05-11 #4379 @kyleniemeyer
@mikemahoney218 Ah, I forgot this was already accepted by rOpenSci (https://github.com/ropensci/software-review/issues/521). In that case, since this has been reviewed as in scope, acceptance in JOSS is fast tracked, and I will both edit and "review".
