### 2022-05-11 #4388 @kyleniemeyer
Ah I see. Yes, please go ahead and archive the software package and let me know the DOI here. The paper looks good, and the review is fast-tracked since it was already reviewed by rOpenSci, so I can go ahead and publish this once I have that.
